+++ src/browser/index.ts
@@ -15,9 +15,9 @@ export { Compiled } from './services/kolay/compiler/reactive.ts';
 export * from './utils.ts';
 
 // Types
+export type { Collection, Manifest, Page } from '../types.ts';
 export type { default as APIDocsService } from './services/kolay/api-docs.ts';
 export type {
   default as DocsService,
   SetupOptions,
 } from './services/kolay/docs.ts';
-export type { Collection, Manifest, Page } from './services/kolay/types.ts';
