--- src/browser/services/kolay/types.ts
@@ -1,22 +1,0 @@
-export interface Manifest {
-  groups: {
-    name: string;
-    list: Page[];
-    tree: Collection;
-  }[];
-}
-
-export interface Collection {
-  path: string;
-  name: string;
-  first?: string;
-  pages: (Collection | Page)[];
-  groupName?: never;
-}
-
-export interface Page {
-  path: string;
-  name: string;
-  groupName: string;
-  cleanedName: string;
-}
